Cleared the structure for easier navigation
