+++ geometry/arc_angle/python/arc_angle.py
@@ -1,0 +1,10 @@
+import math
+
+"""
+    arc_angle
+    @input: radius of circle, length of arc
+    @output: angle of arc
+"""
+def arc_angle(arc_len, radius):
+    angle = (360*float(arc_len))/(math.pi*2*float(radius))
+    return angle
